+++ doc-examples/src/main/java/org/realityforge/arez/doc/examples/inject2/DaggerExample.java
@@ -1,0 +1,10 @@
+package org.realityforge.arez.doc.examples.inject2;
+
+public class DaggerExample
+{
+  public static void main( String[] args )
+  {
+    final ExampleDaggerComponent component = DaggerExampleDaggerComponent.create();
+    component.getExampleConsumer().performSomeAction();
+  }
+}