+++ doc-examples/src/main/java/org/realityforge/arez/doc/examples/inject2/Example.java
@@ -1,0 +1,12 @@
+package org.realityforge.arez.doc.examples.inject2;
+
+import com.google.gwt.core.client.GWT;
+
+public class Example
+{
+  public static void main( String[] args )
+  {
+    final ExampleGinjector injector = GWT.create( ExampleGinjector.class );
+    injector.getExampleConsumer().performSomeAction();
+  }
+}