+++ doc-examples/src/main/java/org/realityforge/arez/doc/examples/inject2/ExampleDaggerComponent.java
@@ -1,0 +1,11 @@
+package org.realityforge.arez.doc.examples.inject2;
+
+import dagger.Component;
+import javax.inject.Singleton;
+
+@Singleton
+@Component( modules = { MyServiceDaggerModule.class, MyEntityRepositoryDaggerModule.class } )
+public interface ExampleDaggerComponent
+{
+  ExampleConsumer getExampleConsumer();
+}