+++ doc-examples/src/main/java/org/realityforge/arez/doc/examples/inject2/ExampleGinjector.java
@@ -1,0 +1,11 @@
+package org.realityforge.arez.doc.examples.inject2;
+
+import com.google.gwt.inject.client.GinModules;
+import com.google.gwt.inject.client.Ginjector;
+
+@GinModules( { ExampleGinModule.class } )
+public interface ExampleGinjector
+  extends Ginjector
+{
+  ExampleConsumer getExampleConsumer();
+}