+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.36](https://github.com/arez/arez/tree/v0.36) (2017-12-13)
+[Full Changelog](https://github.com/arez/arez/compare/v0.35...v0.36)
 
 ##### Added
 * **\[docs\]** Add initial documentation about integrating Arez components into dependency injection